GH-113: Фикс бага, когда падало приложение при удалении желаемой зп с помощью кнопки backspace

+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Filter.kt
@@ -8,6 +8,6 @@ data class Filter(
     val countryId: String? = null,
     val regionId: String? = null,
     val industryId: String? = null,
-    val expectedSalary: Int? = null,
+    val expectedSalary: Long? = null,
     val isOnlyWithSalary: Boolean = false
 )
